Make ClientOpts type safe
This provides some compile-time safety around options passed when
creating a buildkit client.
Before this change since a ClientOpt is an empty interface anything and
everything could be passed in.

This is a breaking change since before consumers could pass in a raw
grpc.DialOption, this now needs to be wrapped with
`client.WithGRPCDialOption(grpcOpt)`
Consumers could also pass in something tat was not a client opt at all
and it was completely ignored.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>